fix(wallet)!: adds wallet signer API (#1601)
Description
---
Adds a local signer API to the wallet SDK
Removed the usage of several key manager calls that return secrets
Removed a couple of key manager calls that return secrets.
fix!: bind the schema version to unsigned transaction signatures

Motivation and Context
---

General-purpose API that signs items that implement the `Signable`
trait.

This trait is implemented for unsigned transaction components, allowing
the API to sign as follows:

```rust
let unsealed_transaction = Transaction::builder()
  //... build tx
    .build();

  let transaction = self.sdk.local_signer_api().sign(
    KeyBranch::Account,
    account.owner_key_id,
     unsealed_transaction,
  )?;
```

The local signer API utilises the locally stored and encrypted Cipher
key; however, the underlying signer is backend-agnostic, meaning that a
cold wallet backend can be implemented. The backend trait is defined as
follows:

```rust
pub trait KeyManagerBackend<M> {
    type Error;

    fn try_sign(&mut self, branch: &str, key_id: KeyId, message: M) -> Result<SignatureOutput, Self::Error>;
}
```

How Has This Been Tested?
---
Manually and existing tests

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - Please specify

BREAKING CHANGE: transactions signatures bind to the schema version (v1)
- previous transaction signatures will be invalid after merge

+++ crates/transaction/src/transaction.rs
@@ -242,7 +242,7 @@ impl Transaction {
         }
     }
 
-    pub const fn schema_version(&self) -> u64 {
+    pub const fn schema_version(&self) -> u16 {
         match self {
             Self::V1(tx) => tx.schema_version(),
         }
